retake

+++ TreeTemperature/Program.cs
@@ -1,4 +1,4 @@
 ﻿
-double a = 36.6;
+float a = 36.6F;
 
 Console.WriteLine($"person temperature: {a}");
